pyproject.toml
[build-system]
requires = ["setuptools >= 75.8.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["gridfm_graphkit*"]
namespaces = false

[project]
name = "gridfm-graphkit"
description = "Grid Foundation Model"
version = "0.0.4"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10,<3.13"

authors = [
  {name = "Matteo Mazzonelli", email = "matteo.mazzonelli1@ibm.com"},
  {name = "Alban Puech", email = "apuech@seas.harvard.edu"},
  {name = "Tamara Govindasamy", email= "tamara.govindasamy@ibm.com"},
  {name = "Mangaliso Mngomezulu", email= "mngomezulum@ibm.com"},
  {name = "Etienne Vos", email= "etienne.vos@ibm.com"},
  {name = "Celia Cintas", email= "celia.cintas@ibm.com"},
  {name = "Jonas Weiss", email= "jwe@zurich.ibm.com"},
]

maintainers = [
  {name = "Matteo Mazzonelli", email = "matteo.mazzonelli1@ibm.com"},
]

keywords = ["electric power grid", "foundational model", "graph neural networks"]

classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence"
]


dependencies = [
    "mlflow>=3.1.0",
    "nbformat>=5.10.4",
    "networkx>=3.4.2",
    "numpy>=2.2.6",
    "pandas>=2.3.0",
    "plotly>=6.1.2",
    "pyyaml>=6.0.2",
    "torch>=2.7.1",
    "torch-geometric>=2.6.1",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
    "lightning",
]

[project.optional-dependencies]
dev = [
  "mkdocs-material",
  "mkdocstrings[python]",
  "pre-commit>=4.2.0",
  "bandit>=1.8.5",
  "build"
]

test = [
  "pytest",
  "pytest-cov"
]

[project.scripts]
gridfm_graphkit = "gridfm_graphkit.__main__:main"
